Cleared the source
