+++ src/main/java/featurecat/lizzie/gui/LizzieFrame.java
@@ -7,19 +7,16 @@ import static java.lang.Math.min;
 
 import com.jhlabs.image.GaussianFilter;
 import featurecat.lizzie.Lizzie;
-import featurecat.lizzie.analysis.GameInfo;
 import featurecat.lizzie.analysis.Leelaz;
 import featurecat.lizzie.analysis.MoveData;
 import featurecat.lizzie.rules.Board;
 import featurecat.lizzie.rules.BoardData;
-import featurecat.lizzie.rules.GIBParser;
 import featurecat.lizzie.rules.SGFParser;
 import featurecat.lizzie.util.Utils;
 import java.awt.*;
 import java.awt.BasicStroke;
 import java.awt.Color;
 import java.awt.Font;
-import java.awt.FontFormatException;
 import java.awt.FontMetrics;
 import java.awt.Frame;
 import java.awt.Graphics;
@@ -44,16 +41,12 @@ import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
 import javax.imageio.ImageIO;
 import javax.swing.*;
-import javax.swing.filechooser.FileNameExtensionFilter;
 import javax.swing.text.html.HTMLDocument;
 import javax.swing.text.html.StyleSheet;
 import org.json.JSONArray;
-import org.json.JSONObject;
 
 /** The window used to display the game. */
 public class LizzieFrame extends MainFrame {
-  private static final ResourceBundle resourceBundle =
-      ResourceBundle.getBundle("l10n.DisplayStrings");
 
   private static final String[] commands = {
     resourceBundle.getString("LizzieFrame.commands.keyN"),
@@ -92,7 +85,6 @@ public class LizzieFrame extends MainFrame {
     resourceBundle.getString("LizzieFrame.commands.keyBackspace"),
     resourceBundle.getString("LizzieFrame.commands.keyE"),
   };
-  private static final String DEFAULT_TITLE = resourceBundle.getString("LizzieFrame.title");
   private static BoardRenderer boardRenderer;
   private static BoardRenderer subBoardRenderer;
   private static VariationTree variationTree;
@@ -105,9 +97,6 @@ public class LizzieFrame extends MainFrame {
 
   private long lastAutosaveTime = System.currentTimeMillis();
   private boolean isReplayVariation = false;
-
-  // Save the player title
-  private String playerTitle = "";
 
   // Display Comment
   private HTMLDocument htmlDoc;
@@ -122,32 +111,11 @@ public class LizzieFrame extends MainFrame {
   // Show the playouts in the title
   private ScheduledExecutorService showPlayouts = Executors.newScheduledThreadPool(1);
   private long lastPlayouts = 0;
-  private String visitsString = "";
   public boolean isDrawVisitsInTitle = true;
-
-  static {
-    // load fonts
-    try {
-      uiFont = new Font("SansSerif", Font.TRUETYPE_FONT, 12);
-      //          Font.createFont(
-      //              Font.TRUETYPE_FONT,
-      //              Thread.currentThread()
-      //                  .getContextClassLoader()
-      //                  .getResourceAsStream("fonts/OpenSans-Regular.ttf"));
-      winrateFont =
-          Font.createFont(
-              Font.TRUETYPE_FONT,
-              Thread.currentThread()
-                  .getContextClassLoader()
-                  .getResourceAsStream("fonts/OpenSans-Semibold.ttf"));
-    } catch (IOException | FontFormatException e) {
-      e.printStackTrace();
-    }
-  }
 
   /** Creates a window */
   public LizzieFrame() {
-    super(DEFAULT_TITLE);
+    super();
 
     boardRenderer = new BoardRenderer(true);
     subBoardRenderer = new BoardRenderer(false);
@@ -281,147 +249,6 @@ public class LizzieFrame extends MainFrame {
   public void clear() {
     if (winrateGraph != null) {
       winrateGraph.clear();
-    }
-  }
-
-  public void openConfigDialog() {
-    ConfigDialog configDialog = new ConfigDialog();
-    configDialog.setVisible(true);
-  }
-
-  public void openChangeMoveDialog() {
-    ChangeMoveDialog changeMoveDialog = new ChangeMoveDialog();
-    changeMoveDialog.setVisible(true);
-  }
-
-  public void toggleGtpConsole() {
-    Lizzie.leelaz.toggleGtpConsole();
-    if (Lizzie.gtpConsole != null) {
-      Lizzie.gtpConsole.setVisible(!Lizzie.gtpConsole.isVisible());
-    } else {
-      Lizzie.gtpConsole = new GtpConsolePane(this);
-      Lizzie.gtpConsole.setVisible(true);
-    }
-  }
-
-  @Override
-  public void startGame() {
-    GameInfo gameInfo = Lizzie.board.getHistory().getGameInfo();
-
-    NewGameDialog gameDialog = new NewGameDialog();
-    gameDialog.setGameInfo(gameInfo);
-    gameDialog.setVisible(true);
-    boolean playerIsBlack = gameDialog.playerIsBlack();
-    boolean isNewGame = gameDialog.isNewGame();
-    //    gameDialog.dispose();
-    if (gameDialog.isCancelled()) return;
-
-    if (isNewGame) {
-      Lizzie.board.clear();
-    }
-    Lizzie.leelaz.komi(gameInfo.getKomi());
-
-    Lizzie.leelaz.time_settings();
-    Lizzie.frame.playerIsBlack = playerIsBlack;
-    Lizzie.frame.isNewGame = isNewGame;
-    Lizzie.frame.isPlayingAgainstLeelaz = true;
-
-    boolean isHandicapGame = gameInfo.getHandicap() != 0;
-    if (isNewGame) {
-      Lizzie.board.getHistory().setGameInfo(gameInfo);
-      if (isHandicapGame) {
-        Lizzie.board.getHistory().getData().blackToPlay = false;
-        Lizzie.leelaz.handicap(gameInfo.getHandicap());
-        if (playerIsBlack) Lizzie.leelaz.genmove("W");
-      } else if (!playerIsBlack) {
-        Lizzie.leelaz.genmove("B");
-      }
-    } else {
-      Lizzie.board.getHistory().setGameInfo(gameInfo);
-      if (Lizzie.frame.playerIsBlack != Lizzie.board.getData().blackToPlay) {
-        if (!Lizzie.leelaz.isThinking) {
-          Lizzie.leelaz.genmove((Lizzie.board.getData().blackToPlay ? "B" : "W"));
-        }
-      }
-    }
-  }
-
-  public void editGameInfo() {
-    GameInfo gameInfo = Lizzie.board.getHistory().getGameInfo();
-
-    GameInfoDialog gameInfoDialog = new GameInfoDialog();
-    gameInfoDialog.setGameInfo(gameInfo);
-    gameInfoDialog.setVisible(true);
-
-    gameInfoDialog.dispose();
-  }
-
-  public void saveFile() {
-    FileNameExtensionFilter filter = new FileNameExtensionFilter("*.sgf", "SGF");
-    JSONObject filesystem = Lizzie.config.persisted.getJSONObject("filesystem");
-    JFileChooser chooser = new JFileChooser(filesystem.getString("last-folder"));
-    chooser.setFileFilter(filter);
-    chooser.setMultiSelectionEnabled(false);
-    int result = chooser.showSaveDialog(null);
-    if (result == JFileChooser.APPROVE_OPTION) {
-      File file = chooser.getSelectedFile();
-      if (file.exists()) {
-        int ret =
-            JOptionPane.showConfirmDialog(
-                null,
-                resourceBundle.getString("LizzieFrame.prompt.sgfExists"),
-                "Warning",
-                JOptionPane.OK_CANCEL_OPTION);
-        if (ret == JOptionPane.CANCEL_OPTION) {
-          return;
-        }
-      }
-      if (!file.getPath().endsWith(".sgf")) {
-        file = new File(file.getPath() + ".sgf");
-      }
-      try {
-        SGFParser.save(Lizzie.board, file.getPath());
-        filesystem.put("last-folder", file.getParent());
-      } catch (IOException err) {
-        JOptionPane.showConfirmDialog(
-            null,
-            resourceBundle.getString("LizzieFrame.prompt.failedTosaveFile"),
-            "Error",
-            JOptionPane.ERROR);
-      }
-    }
-  }
-
-  public void openFile() {
-    FileNameExtensionFilter filter = new FileNameExtensionFilter("*.sgf or *.gib", "SGF", "GIB");
-    JSONObject filesystem = Lizzie.config.persisted.getJSONObject("filesystem");
-    JFileChooser chooser = new JFileChooser(filesystem.getString("last-folder"));
-
-    chooser.setFileFilter(filter);
-    chooser.setMultiSelectionEnabled(false);
-    int result = chooser.showOpenDialog(null);
-    if (result == JFileChooser.APPROVE_OPTION) loadFile(chooser.getSelectedFile());
-  }
-
-  public void loadFile(File file) {
-    JSONObject filesystem = Lizzie.config.persisted.getJSONObject("filesystem");
-    if (!(file.getPath().endsWith(".sgf") || file.getPath().endsWith(".gib"))) {
-      file = new File(file.getPath() + ".sgf");
-    }
-    try {
-      System.out.println(file.getPath());
-      if (file.getPath().endsWith(".sgf")) {
-        SGFParser.load(file.getPath());
-      } else {
-        GIBParser.load(file.getPath());
-      }
-      filesystem.put("last-folder", file.getParent());
-    } catch (IOException err) {
-      JOptionPane.showConfirmDialog(
-          null,
-          resourceBundle.getString("LizzieFrame.prompt.failedToOpenFile"),
-          "Error",
-          JOptionPane.ERROR);
     }
   }
 
@@ -1438,19 +1265,6 @@ public class LizzieFrame extends MainFrame {
     }
   }
 
-  public void setPlayers(String whitePlayer, String blackPlayer) {
-    playerTitle = String.format("(%s [W] vs %s [B])", whitePlayer, blackPlayer);
-    updateTitle();
-  }
-
-  public void updateTitle() {
-    StringBuilder sb = new StringBuilder(DEFAULT_TITLE);
-    sb.append(playerTitle);
-    sb.append(" [" + Lizzie.leelaz.engineCommand() + "]");
-    sb.append(visitsString);
-    setTitle(sb.toString());
-  }
-
   private void setDisplayedBranchLength(int n) {
     boardRenderer.setDisplayedBranchLength(n);
   }
@@ -1467,11 +1281,6 @@ public class LizzieFrame extends MainFrame {
 
   public boolean incrementDisplayedBranchLength(int n) {
     return boardRenderer.incrementDisplayedBranchLength(n);
-  }
-
-  public void resetTitle() {
-    playerTitle = "";
-    updateTitle();
   }
 
   public void copySgf() {
