+++ src/main/java/featurecat/lizzie/gui/NewGameDialog.java
@@ -26,8 +26,7 @@ import javax.swing.border.EmptyBorder;
 
 /** @author unknown */
 public class NewGameDialog extends JDialog {
-  private static final ResourceBundle resourceBundle =
-      ResourceBundle.getBundle("l10n.DisplayStrings");
+  private static final ResourceBundle resourceBundle = MainFrame.resourceBundle;
   // create formatters
   public static final DecimalFormat FORMAT_KOMI = new DecimalFormat("#0.0");
   public static final DecimalFormat FORMAT_HANDICAP = new DecimalFormat("0");
